# Byxis : Passage à la version 5.0
## Public patch note
- Réglage de certains textes, qui contenait des espaces en trop.
- Les durées de craft peuvent désormais être personnalisées pour chaque items. Ce qui permettra la réduction de temps craft de munitions par exemple.
- Ajout des médicaments, permettant le craft d'item de soin.
- Réglage de l'invincibilité des zombies villageois
- Les joueurs infectés peuvent désormais contaminer les autres en les tapants à la main

## Private patch note
- Ajout du support PostgreSQL en bdd
- Fix de bugs/clarité de code sur le workshop
- Passage du projet en UTF8 (à peut être causé des problèmes de nom, qui seront corrigé avec le temps

+++ src/main/java/fr/byxis/faction/zone/DataZone.java
@@ -155,7 +155,7 @@ public class DataZone {
                 final PreparedStatement ps = connection.prepareStatement("INSERT INTO faction_zone(faction_name,zone,duration) VALUES (?,?,?)");
                 ps.setString(2, zone);
                 ps.setString(1, factionName);
-                ps.setInt(3, (int) diff);
+                ps.setLong(3, diff);
                 ps.executeUpdate();
             }
         } catch (SQLException e) {
